# Inline Index::as_usize()