@adevore 2017-07-14
Interesting, I find that surprising. Thanks for taking a look.