@adevore 2017-07-14
I did some further benchmarks. This improves benchmark times across nearly all non-trivial benchmarks. It's a good win.